refactor: make window scrollable
- make window scrollable
- move save/load preset logic to presets.py
- add Alt+F4 keybind to Close action
